[refactoring] find winner를 cars 메소드로 변경, 패키지 변겅 수정
